Move LogEvents into custom hook
