+++ examples/cra-dapp/src/hooks/use-log-events.ts
@@ -1,0 +1,26 @@
+import { useEffect } from 'react';
+import { useFuel } from '@fuel-wallet/react';
+
+export function useLogEvents() {
+  const { fuel } = useFuel();
+
+  useEffect(() => {
+    const log = (prefix: string) => (data: any) => {
+      console.log(prefix, data);
+    };
+    const logAccounts = log('accounts');
+    const logConnection = log('connection');
+    const logCurrentAccount = log('currentAccount');
+
+    fuel.on(fuel.events.accounts, logAccounts);
+    fuel.on(fuel.events.connection, logConnection);
+    fuel.on(fuel.events.currentAccount, logCurrentAccount);
+    return () => {
+      fuel.off(fuel.events.accounts, logAccounts);
+      fuel.off(fuel.events.connection, logConnection);
+      fuel.off(fuel.events.currentAccount, logCurrentAccount);
+    };
+  }, [fuel]);
+
+  return null;
+}
